% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_config.R
\name{build_config}
\alias{build_config}
\title{Build config file for rLHFA}
\usage{
build_config(LakeName, directory)
}
\arguments{
\item{LakeName}{Name of the lake}

\item{directory}{Folder where the input files are stored}
}
\description{
Build config file for rLHFA
}
