% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lwnet.R
\name{calc_lwnet}
\alias{calc_lwnet}
\title{Calculate net long wave radiation}
\arguments{
\item{time}{Vector of times as POSIXct}

\item{lat}{latitude for which to calculate clear sky radiation}

\item{press}{Air pressure in (hPa) if available}

\item{ta}{Air temperature in (°C) if available}

\item{rh}{relative humidity in percentage if available}

\item{sw}{observed showrt wave radiation}

\item{ts}{Water surface temperature in °C}
}
\value{
A data.frame with lw, LWo, and lwnet
}
\description{
Calculate net long wave radiation
}
\references{
Read et al. 2012
}
