% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownSample_TS.R
\name{DownSample_TS}
\alias{DownSample_TS}
\title{Down sample time series}
\usage{
DownSample_TS(Dates, outRs, VarArray = NULL)
}
\arguments{
\item{Dates}{Vector of original data times}

\item{outRs}{Sampling inerval for output time steps}

\item{VarArray}{Array of original data}
}
\description{
Down sample time series
}
