% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmithGamma.R
\name{getSmithGamma}
\alias{getSmithGamma}
\title{Function to get Smith Gamma}
\usage{
getSmithGamma(lat, time)
}
\arguments{
\item{lat}{Latitude}

\item{time}{time as POSIXct format}
}
\value{
Smiths Gamma
}
\description{
Function to get Smith Gamma
}
\examples{
 lat <- 54
 time <- as.POSIXct("2001-10-10",tz="UTC")
 getSmithGamma(lat,time)
}
