% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFileOpen.R
\name{gFileOpen}
\alias{gFileOpen}
\title{read in forcing data for heat flux calculations}
\usage{
gFileOpen(fileName, treatAsWtr = FALSE)
}
\arguments{
\item{fileName}{Name of the file to read in}

\item{treadAsWtr}{Logical, is the inputfile water temperature with several depths?}
}
\value{
Returns a list with date, data and if treatAsWtr = TRUE depths of the observations
}
\description{
read in forcing data for heat flux calculations
}
\author{
Johannes Feldbauer
}
