% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mdl_LW.R
\name{Mdl_LW}
\alias{Mdl_LW}
\title{Model long wave radiation}
\usage{
Mdl_LW(airT, RH, clf, month)
}
\arguments{
\item{airT}{air temperature in degrees C}

\item{RH}{relative humidity in \%}

\item{clf}{cloud cover fraction}

\item{month}{month of the year}
}
\value{
numerical vector of long wave radiation
}
\description{
Model long wave radiation
}
\examples{
Mdl_LW(20,50,0.1,12)
}
