% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanmean.R
\name{nanmean}
\alias{nanmean}
\title{calculate mean of a vector containing NaNs}
\usage{
nanmean(inputdata)
}
\arguments{
\item{inputdata}{data for which to calculate the mean}
}
\value{
arithmetic mean of inputdata with NaNs removed. Same as mean(x,na.rm=TRUE)
}
\description{
calculate mean of a vector containing NaNs
}
\examples{
x <- c(1,3,1,4,NaN,4)
nanmean(x)
}
\seealso{
mean()
}
