% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenCfg.R
\name{OpenCfg}
\alias{OpenCfg}
\title{function to read in config file for rLHFA}
\usage{
OpenCfg(LakeName, folder)
}
\arguments{
\item{LakeName}{Name of the Lake. Forcing and config file must have the same name}

\item{folder}{Folder where the config file is stored}
}
\value{
Returns a list with options from the config file
}
\description{
function to read in config file for rLHFA
}
