% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatVaporFromTemp.R
\name{SatVaporFromTemp}
\alias{SatVaporFromTemp}
\title{Calculate saturation vapor pressure from temperature}
\usage{
SatVaporFromTemp(Temperature)
}
\arguments{
\item{Temperature}{Water temperature}
}
\value{
Vector of saturation vapor pressures in bar
}
\description{
Calculate saturation vapor pressure from temperature
}
\examples{
SatVaporFromTemp(15)
}
\references{
Empirical fit from P.R. Lowe, 1976. An Approximating Polynomial for the
Computation of Saturation Vapor Pressure. Journal of Applied Meteorology
}
