% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_latent.R
\name{sens_latent}
\alias{sens_latent}
\title{Calculate sensible and latent heat flux}
\source{
Zeng, X., Zhao, M., Dickson, R.E. 1998. Intercomparison of Bulk
 aerodynamic algorithms for the computation of sea surface fluxes using
 TOGA COARE and TAO data. Journal of Climate 11: 2628-2644.
}
\usage{
sens_latent(ts, Uz, ta, rh, hu, ht, hq, alt, lat)
}
\arguments{
\item{ts}{surface temperature, degrees C.}

\item{Uz}{wind speed, m/s.}

\item{ta}{air temperature, degrees C.}

\item{rh}{relative humidity, %.}

\item{hu}{height of wind measurement, m.}

\item{ht}{height of air temperature measurement, m.}

\item{hq}{heigh of humidity measurement, m.}

\item{alt}{altitude of lake, m.}
}
\description{
Calculate sensible and latent heat flux
}
