% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputConstructor.R
\name{OutputConstructor}
\alias{OutputConstructor}
\title{Construct input data dependency}
\usage{
OutputConstructor(outputNames, pltMods)
}
\arguments{
\item{outputNames}{Names of the variables that should be outputed}

\item{pltMods}{Modifications to the plot settings}
}
\value{
Returns a list with several tables and lists for further calculations
}
\description{
Construct input data dependency
}
