% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaporPressure.R
\name{VaporPressure}
\alias{VaporPressure}
\title{calculate vapor pressure from relative humidity}
\usage{
VaporPressure(Temperature, RelativeHumidity)
}
\arguments{
\item{Temperature}{Air temperature in degrees C}

\item{RelativeHumidity}{Relative humidity as 0-1}
}
\value{
Numerical vector of vapor pressures
}
\description{
calculate vapor pressure from relative humidity
}
